requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai>=1.30.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-multipart==0.0.6